+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * Improve parameter type naming for generic types (#343)
 
 ## Bug fixes:
+* Fix: Replace EoL Specflow.Internal.Json with System.Text.Json. This may result in slightly different json parsing for reqnroll.json.
 
 *Contributors of this release (in alphabetical order):* @clrudolphi, @obligaron, @olegKoshmeliuk